orchestrator-bundle/orc8r-libs/requirements.txt
git+https://github.com/canonical/operator/#egg=ops
ops-lib-pgsql
lightkube
lightkube-models
